+++ changelog.md
@@ -14,7 +14,7 @@ SortEdgeIndices(in_place=True)(GraphDict())
 * Fixed error in ``ClinToxDataset`` label index.
 * Reworked ``kgcnn.graph.adj.get_angle_index`` with additional function arguments. Default behaviour remains identical. For periodic system an additional `allow_reverse_edges=True` is now required.
 * Added input embedding for edges in ``kgcnn.literature.EGNN`` and changed order of input tensors. Debugged model.
-
+* Reworked ``kgcnn.literature.PAiNN`` to simplify normalization option and add equivariant initialization method.
 
 v2.1.0